+++ Swiftly/SmtpManager.cpp
@@ -33,7 +33,7 @@ void SmtpManager::sendMail(const QString &from, const QString &to, const QString
     this->from = from;
     rcpt = to;
     state = Init;
-    socket->connectToHostEncrypted(host, port); //"smtp.gmail.com" and 465 for gmail TLS
+    socket->connectToHostEncrypted(host, static_cast<quint16>(port)); //"smtp.gmail.com" and 465 for gmail TLS
     if (!socket->waitForConnected(timeout)) {
          qDebug() << socket->errorString();
      }
